agent/requirements.txt
langgraph==0.2.33
openai==1.55.3
python-dotenv==1.0.1
requests==2.32.3
